move predicate-related code into pyramid.config.predicates
