+++ src/pyramid/config/i18n.py
@@ -3,7 +3,7 @@ from pyramid.interfaces import ILocaleNegotiator, ITranslationDirectories
 from pyramid.exceptions import ConfigurationError
 from pyramid.path import AssetResolver
 
-from pyramid.config.util import action_method
+from pyramid.config.actions import action_method
 
 
 class I18NConfiguratorMixin(object):
