+++ src/pyramid/config/rendering.py
@@ -1,7 +1,7 @@
 from pyramid.interfaces import IRendererFactory, PHASE1_CONFIG
 
 from pyramid import renderers
-from pyramid.config.util import action_method
+from pyramid.config.actions import action_method
 
 DEFAULT_RENDERERS = (
     ('json', renderers.json_renderer_factory),
