+++ tests/test_config/test_actions.py
@@ -50,7 +50,7 @@ class ConfiguratorTests(unittest.TestCase):
         self.assertEqual(config.action('discrim', kw={'a': 1}), None)
 
     def test_action_autocommit_with_introspectables(self):
-        from pyramid.config.util import ActionInfo
+        from pyramid.config.actions import ActionInfo
 
         config = self._makeOne(autocommit=True)
         intr = DummyIntrospectable()
@@ -1027,6 +1027,41 @@ class Test_resolveConflicts(unittest.TestCase):
             ]
         )
         self.assertRaises(ConfigurationConflictError, list, result)
+
+
+class TestActionInfo(unittest.TestCase):
+    def _getTargetClass(self):
+        from pyramid.config.actions import ActionInfo
+
+        return ActionInfo
+
+    def _makeOne(self, filename, lineno, function, linerepr):
+        return self._getTargetClass()(filename, lineno, function, linerepr)
+
+    def test_class_conforms(self):
+        from zope.interface.verify import verifyClass
+        from pyramid.interfaces import IActionInfo
+
+        verifyClass(IActionInfo, self._getTargetClass())
+
+    def test_instance_conforms(self):
+        from zope.interface.verify import verifyObject
+        from pyramid.interfaces import IActionInfo
+
+        verifyObject(IActionInfo, self._makeOne('f', 0, 'f', 'f'))
+
+    def test_ctor(self):
+        inst = self._makeOne('filename', 10, 'function', 'src')
+        self.assertEqual(inst.file, 'filename')
+        self.assertEqual(inst.line, 10)
+        self.assertEqual(inst.function, 'function')
+        self.assertEqual(inst.src, 'src')
+
+    def test___str__(self):
+        inst = self._makeOne('filename', 0, 'function', '   linerepr  ')
+        self.assertEqual(
+            str(inst), "Line 0 of file filename:\n       linerepr  "
+        )
 
 
 def _conflictFunctions(e):
